refactor: separate ItemRow from matches component for better modularity
